Rename `secondary` folder to `common`
