+++ src/types/Story.ts
@@ -3,7 +3,7 @@ import Entity from './Entity';
 import Culture from './Culture';
 import { NewsroomRef } from './Newsroom';
 import UserAccountRef from './UserAccountRef';
-import { OEmbedInfo } from './secondary';
+import { OEmbedInfo } from './common';
 
 export enum FormatVersion {
     HTML = 1,
